+++ _version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "4.0.41";
+export declare const version = "4.0.42";