+++ utils/address.d.ts
@@ -6,3 +6,10 @@ export declare function getContractAddress(transaction: {
     from: string;
     nonce: Arrayish | BigNumber | number;
 }): string;
+export declare type Create2Options = {
+    from: string;
+    salt: Arrayish;
+    initCode?: Arrayish;
+    initCodeHash?: Arrayish;
+};
+export declare function getCreate2Address(options: Create2Options): string;